Time: 0 ms (100.00%) | Memory: 8 MB (41.33%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -3,7 +3,7 @@ public:
     int findComplement(int num) {
        int n=num;
        int c=0;
-       long long r=0;
+       long  r=0;
        long  power=1;
        while(n!=0)
        {
